Experiments.sampledata moved to dedicated data loading class
